requirements.txt
alembic==1.4.1
antlr4-python3-runtime==4.8
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cryptography==3.2.1
cycler==0.10.0
databricks-cli==0.14.1
docker==4.4.0
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
hydra-core==1.0.4
idna==2.10
importlib-resources==3.3.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kaggle==1.5.10
kiwisolver==1.3.1
llvmlite==0.35.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mlflow==1.12.1
msrest==0.6.19
numba==0.52.0
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.5
pandas==1.1.4
Pillow==8.0.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.14.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.4
shap==0.37.0
six==1.15.0
sklearn==0.0
slicer==0.0.3
smmap==3.0.4
SQLAlchemy==1.3.20
sqlparse==0.4.1
tabulate==0.8.7
text-unidecode==1.3
threadpoolctl==2.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
